Extract dialog methods to class
