+++ app/src/androidTest/java/com/akjaw/timicompose/stopwatch/StopwatchScreenVerifier.kt
@@ -2,12 +2,10 @@ package com.akjaw.timicompose.stopwatch
 
 import androidx.compose.ui.test.SemanticsNodeInteraction
 import androidx.compose.ui.test.assert
-import androidx.compose.ui.test.filter
 import androidx.compose.ui.test.filterToOne
 import androidx.compose.ui.test.hasAnyChild
 import androidx.compose.ui.test.hasAnyDescendant
 import androidx.compose.ui.test.hasText
-import androidx.compose.ui.test.onAllNodesWithTag
 import androidx.compose.ui.test.onChildren
 import androidx.compose.ui.test.onNodeWithTag
 import com.akjaw.timicompose.ActivityComposeTestRule
@@ -33,28 +31,6 @@ class StopwatchScreenVerifier(
             .get(index)
 
         addButton.assert(hasAnyChild(hasText("Add a new stopwatch")))
-    }
-
-    // TODO extract to dialogVerifier
-    fun confirmTaskListHasSize(amount: Int) {
-        val nodes = composeTestRule
-            .onAllNodesWithTag("AddStopwatchTaskItem")
-            .fetchSemanticsNodes()
-
-        assert(nodes.count() == amount) {
-            "Expected $amount tasks but got ${nodes.count()}"
-        }
-    }
-
-    // TODO extract to dialogVerifier
-    fun confirmTaskListDoesNotContain(taskName: String) {
-        val nodes = composeTestRule
-            .onAllNodesWithTag("AddStopwatchTaskItem")
-            .filter(hasAnyDescendant(hasText(taskName))) // TODO extract with assertion
-
-        assert(nodes.fetchSemanticsNodes(atLeastOneRootRequired = false).isEmpty()) {
-            "The task with the name: $taskName exists but shouldn't"
-        }
     }
 
     private fun selectStopwatchWithTaskName(name: String): SemanticsNodeInteraction =
